Move releases confirm into separate component
